requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
cloudinary==1.36.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-admin-tailwind==0.0.31
django-admin-volt==1.0.10
django-appconf==1.0.6
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-compressor==4.4
django-cors-headers==4.3.0
django-embed-video==1.4.9
django-environ==0.11.2
django-import-export==3.3.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-parler==2.3
django-taggit==4.0.0
django-unfold==0.17.1
djangoql==0.17.1
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
fontawesomefree==6.5.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.1
ply==3.11
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-gettext==5.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
xlrd==2.0.1
xlwt==1.3.0
